+++ packages/dartcv/test/imgproc/imgproc_async_test.dart
@@ -254,9 +254,20 @@ void main() async {
   });
 
   test('cv.thresholdAsync', () async {
-    final src = await cv.imreadAsync("test/images/circles.jpg", flags: cv.IMREAD_COLOR);
-    final (_, dst) = await cv.thresholdAsync(src, 100, 255, cv.THRESH_BINARY);
-    expect((dst.width, dst.height, dst.channels), (src.width, src.height, src.channels));
+    {
+      final src = await cv.imreadAsync("test/images/circles.jpg", flags: cv.IMREAD_COLOR);
+      final (_, dst) = await cv.thresholdAsync(src, 100, 255, cv.THRESH_BINARY);
+      expect((dst.width, dst.height, dst.channels), (src.width, src.height, src.channels));
+    }
+
+    // threshold with mask
+    {
+      final src = await cv.imreadAsync("test/images/circles.jpg", flags: cv.IMREAD_COLOR);
+      final mask = cv.Mat.ones(src.rows, src.cols, cv.MatType.CV_8UC1);
+      final (_, dst) = await cv.thresholdAsync(src, 100, 255, cv.THRESH_BINARY, mask: mask);
+      // cv.imwrite("test/images_out/threshold_mask.png", dst);
+      expect((dst.width, dst.height, dst.channels), (src.width, src.height, src.channels));
+    }
   });
 
   test('cv.distanceTransformAsync', () async {