+++ packages/dartcv/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.7
+
+* add `Mat.reinterpret`
+* add `fisheye.solvePnPRansac`, support `mask` for `cv.threshold`
+* add `StereoBM`, `StereoSGBM` from `calib3d` module
+
 ## 1.1.6
 
 * add `FlannIndexParams`, `FlannSearchParams`, `FlannKDTreeIndexParams`